extract data-handling methods from optimize
